+++ golem-worker-service-base/src/api/custom_http_request_api.rs
@@ -1,17 +1,16 @@
 use std::sync::Arc;
 
-use crate::http_api_definition::ResponseMapping;
 use async_trait::async_trait;
 use hyper::header::HOST;
 use poem::http::StatusCode;
 use poem::{Body, Endpoint, Request, Response};
 use tracing::{error, info};
+use crate::api::register_api_definition_api::ResponseMapping;
+use crate::http::http_request::{ApiInputPath, InputHttpRequest};
 
-use crate::worker_binding_resolver::WorkerBindingResolver;
-use crate::http_request::{ApiInputPath, InputHttpRequest};
 use crate::service::http_request_definition_lookup::HttpRequestDefinitionLookup;
+use crate::worker_request::worker_request_to_response::WorkerRequestToResponse;
 use crate::worker_request::WorkerRequest;
-use crate::worker_request_to_response::WorkerRequestToResponse;
 
 // Executes custom request with the help of worker_request_executor and definition_service
 // This is a common API projects can make use of, similar to healthcheck service
