+++ golem-worker-service-base/src/http/mod.rs
@@ -1,4 +1,3 @@
 pub mod http_request;
 pub mod http_api_definition;
-pub mod http_worker_binding_resolver;
 pub mod oas_worker_bridge;
