+++ golem-worker-service-base/src/parser/literal_parser.rs
@@ -1,9 +1,9 @@
 use nom::character::complete::not_line_ending;
 use nom::combinator::all_consuming;
 use nom::IResult;
+use crate::http::http_api_definition::LiteralInfo;
 
 use super::*;
-use crate::http_api_definition::LiteralInfo;
 
 pub struct LiteralParser;
