+++ golem-worker-service-base/src/service/http_request_definition_lookup.rs
@@ -1,7 +1,7 @@
-use crate::http_api_definition::HttpApiDefinition;
-use crate::http_request::InputHttpRequest;
 use async_trait::async_trait;
 use std::fmt::Display;
+use crate::http::http_api_definition::HttpApiDefinition;
+use crate::http::http_request::InputHttpRequest;
 
 #[async_trait]
 pub trait HttpRequestDefinitionLookup {
